fix(web): move friendlyName out of MessageBubble.tsx to fix CI test failure
tool-explain.ts (a pure string-formatting utility, no React) imported
friendlyName from MessageBubble.tsx (a React component), pulling the whole
React dependency chain into its module graph. This passed locally (web/
node_modules happened to be installed) but failed in CI, which only runs
`npm ci` in the root turbollm/ package: `Cannot find package 'react'`.

Moved friendlyName into tool-explain.ts (the natural home — it's already
the shared "describe a tool call" utility) and updated the three importers
to pull it from there instead, reversing the dependency direction so a
utility module never depends on a component file. Also removes the
tool-explain.test.ts localStorage stub, which was only needed to survive
this same bad import chain — verified by running the test with web/
node_modules removed entirely (reproducing the CI failure), confirming it
now passes standalone.
