+++ turbollm/web/src/lib/tool-explain.ts
@@ -1,4 +1,9 @@
-import { friendlyName } from '../screens/chat/MessageBubble'
+/** Strips an MCP tool's server-prefix (`mcp__server__tool` → `tool`) and turns
+ *  underscores into spaces for display, e.g. in the approval bar and tool-call cards. */
+export function friendlyName(name: string): string {
+  if (name.startsWith('mcp__')) return name.replace(/^mcp__[^_]+(?:_[^_]+)*__/, '')
+  return name.replace(/_/g, ' ')
+}
 
 /** PURE: mirrors resolveSearchQuery in src/tools/builtin.ts — the schema declares a single
  *  required `query: string`, but a model sometimes emits `queries: string[]` instead (seen from
