+++ turbollm/web/src/screens/settings/ToolPermissionsSection.tsx
@@ -4,7 +4,7 @@ import { ChevronRight, ShieldCheck } from 'lucide-react'
 import { useSettings } from '../../lib/queries'
 import { ApiError, type ToolPolicy } from '../../lib/api'
 import { fetchAvailableTools } from '../../lib/chat-api'
-import { friendlyName } from '../chat/MessageBubble'
+import { friendlyName } from '../../lib/tool-explain'
 import { Collapsible, CollapsibleContent, CollapsibleTrigger } from '../../components/ui/collapsible'
 import { toast } from '../../components/ui/sonner'
 import { cn } from '../../lib/utils'
